pyproject.toml
[project]
name = "reachy-mini-sdk-skill"
version = "2.2.0"
description = "Comprehensive Claude Skill for programming Reachy Mini robots with Python SDK and REST API"
authors = [{ name = "Jean-Jacques Martres", email = "jjmartres@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
keywords = [
    "reachy-mini",
    "robotics",
    "claude-ai",
    "opencode",
    "anthropic",
    "openapi",
    "sdk",
    "rest-api",
    "ai-integration"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Hardware",
]

dependencies = [
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pytest>=7.0.0",
]

[project.urls]
Homepage = "https://github.com/jjmartres/reachy-mini-sdk-skill"
Documentation = "https://github.com/jjmartres/reachy-mini-sdk-skill#readme"
Repository = "https://github.com/jjmartres/reachy-mini-sdk-skill.git"
Issues = "https://github.com/jjmartres/reachy-mini-sdk-skill/issues"
Changelog = "https://github.com/jjmartres/reachy-mini-sdk-skill/releases"

[project.scripts]
package-skill = "scripts.package_skill:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.tox
  | build
  | dist
  | __pycache__
)/
'''

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = "yaml.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
